# shared/schemas/openapi.yaml
openapi: 3.0.0
info:
  title: AI Ad Intelligence & Creation Suite API
  version: 1.0.0
  description: API for video ad intelligence, creation, and publishing
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://gateway-api-{project}.run.app
    description: Cloud Run production
paths:
...
  /render/remix:
    post:
      summary: Create a new video remix from storyboard
      tags:
        - Render
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemixRequest'
      responses:
        '202':
          description: Remix job created
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId:
                    type: string
                  status:
                    type: string
  /render/jobs/{id}:
    get:
      summary: Get status of a render job
      tags:
        - Render
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenderJob'
...
components:
  schemas:
...
    RemixRequest:
      type: object
      required:
        - storyboard
        - assetMap
      properties:
        storyboard:
          type: array
          items:
            $ref: '#/components/schemas/StoryboardClip'
        assetMap:
          type: object
        outputFormat:
          type: string
        overlays:
          type: object
    StoryboardClip:
      type: object
      required:
        - clipId
        - duration
      properties:
        clipId:
          type: string
        duration:
          type: number
        transition:
          type: string
        effects:
          type: array
          items:
            type: string
    RenderJob:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed]
        progress:
          type: number
        outputUrl:
          type: string
        error:
          type: string
        createdAt:
          type: string
        completedAt:
          type: string
...